+++ ClassLibrary/I3V3.LogEvents/RecCallStartLogEvent.cs
@@ -6,7 +6,7 @@
 namespace I3V3.LogEvents
 {
     /// <summary>
-    /// Data class for the RecCallStartLogEvent. See Sections 4.12.3.7 and E.8.1 of NENA-STA-010.3.
+    /// Data class for the RecCallStartLogEvent. See Sections 4.12.3.7 and E.8.1 of NENA-STA-010.3b.
     /// Logged when a SIPREC media recorder detects the start of a SIPREC call.
     /// </summary>
     public class RecCallStartLogEvent : CallLogEvent